refactor: extract highlightFileByLines and escapeHtml to shared module
Move escapeHtml, hastToHtml, and highlightFileByLines from the duplicated
implementations in src/api/diff.ts and src/browser/lib/diff-worker.ts into
src/shared/diff-utils.ts. Also remove the local escapeHtml copy in
src/browser/lib/interdiff.ts.

Closes #21
